+++ test/Application/TestAsset/InteropMiddleware.php
@@ -7,9 +7,9 @@
 
 namespace ZendTest\Expressive\Application\TestAsset;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as DelegateInterface;
+use Webimpress\HttpMiddlewareCompatibility\MiddlewareInterface;
 
 class InteropMiddleware implements MiddlewareInterface
 {