+++ test/ApplicationTest.php
@@ -9,7 +9,6 @@ namespace ZendTest\Expressive;
 
 use DomainException;
 use Fig\Http\Message\StatusCodeInterface as StatusCode;
-use Interop\Http\ServerMiddleware\DelegateInterface;
 use InvalidArgumentException;
 use Mockery;
 use PHPUnit\Framework\TestCase;
@@ -21,6 +20,7 @@ use Psr\Http\Message\ServerRequestInterface;
 use ReflectionProperty;
 use RuntimeException;
 use UnexpectedValueException;
+use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as DelegateInterface;
 use Zend\Diactoros\Response\EmitterInterface;
 use Zend\Diactoros\Response\SapiEmitter;
 use Zend\Diactoros\ServerRequest;
@@ -39,6 +39,8 @@ use Zend\Expressive\Router\RouterInterface;
 use Zend\Expressive\Template\TemplateRendererInterface;
 use Zend\Stratigility\MiddlewarePipe;
 use Zend\Stratigility\Route as StratigilityRoute;
+
+use const Webimpress\HttpMiddlewareCompatibility\HANDLER_METHOD;
 
 /**
  * @covers Zend\Expressive\Application
@@ -294,7 +296,7 @@ class ApplicationTest extends TestCase
 
         $finalResponse = $this->prophesize(ResponseInterface::class)->reveal();
         $defaultDelegate = $this->prophesize(DelegateInterface::class);
-        $defaultDelegate->process(Argument::type(ServerRequestInterface::class))
+        $defaultDelegate->{HANDLER_METHOD}(Argument::type(ServerRequestInterface::class))
             ->willReturn($finalResponse);
 
         $emitter = $this->prophesize(EmitterInterface::class);
@@ -338,7 +340,7 @@ class ApplicationTest extends TestCase
 
         $finalResponse = $this->prophesize(ResponseInterface::class)->reveal();
         $defaultDelegate = $this->prophesize(DelegateInterface::class);
-        $defaultDelegate->process(Argument::type(ServerRequestInterface::class))
+        $defaultDelegate->{HANDLER_METHOD}(Argument::type(ServerRequestInterface::class))
             ->willReturn($finalResponse);
 
         $emitter = $this->prophesize(EmitterInterface::class);