+++ test/Middleware/ImplicitHeadMiddlewareTest.php
@@ -9,16 +9,18 @@ namespace ZendTest\Expressive\Middleware;
 
 use Fig\Http\Message\RequestMethodInterface as RequestMethod;
 use Fig\Http\Message\StatusCodeInterface as StatusCode;
-use Interop\Http\ServerMiddleware\DelegateInterface;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
+use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as DelegateInterface;
 use Zend\Diactoros\Response;
 use Zend\Expressive\Middleware\ImplicitHeadMiddleware;
 use Zend\Expressive\Router\Route;
 use Zend\Expressive\Router\RouteResult;
+
+use const Webimpress\HttpMiddlewareCompatibility\HANDLER_METHOD;
 
 class ImplicitHeadMiddlewareTest extends TestCase
 {
@@ -30,7 +32,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())
+        $delegate->{HANDLER_METHOD}($request->reveal())
             ->willReturn($response);
 
         $middleware = new ImplicitHeadMiddleware();
@@ -48,7 +50,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())
+        $delegate->{HANDLER_METHOD}($request->reveal())
             ->willReturn($response);
 
         $middleware = new ImplicitHeadMiddleware();
@@ -69,7 +71,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())
+        $delegate->{HANDLER_METHOD}($request->reveal())
             ->willReturn($response);
 
         $middleware = new ImplicitHeadMiddleware();
@@ -93,7 +95,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())
+        $delegate->{HANDLER_METHOD}($request->reveal())
             ->willReturn($response);
 
         $middleware = new ImplicitHeadMiddleware();
@@ -118,7 +120,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())->shouldNotBeCalled($response);
+        $delegate->{HANDLER_METHOD}($request->reveal())->shouldNotBeCalled($response);
 
         $middleware = new ImplicitHeadMiddleware();
         $result = $middleware->process($request->reveal(), $delegate->reveal());
@@ -144,7 +146,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
         $response = $this->prophesize(ResponseInterface::class)->reveal();
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())->shouldNotBeCalled($response);
+        $delegate->{HANDLER_METHOD}($request->reveal())->shouldNotBeCalled($response);
 
         $expected   = new Response();
         $middleware = new ImplicitHeadMiddleware($expected);
@@ -177,7 +179,7 @@ class ImplicitHeadMiddlewareTest extends TestCase
             ->will([$response, 'reveal']);
 
         $delegate = $this->prophesize(DelegateInterface::class);
-        $delegate->process($request->reveal())->will([$response, 'reveal']);
+        $delegate->{HANDLER_METHOD}($request->reveal())->will([$response, 'reveal']);
 
         $middleware = new ImplicitHeadMiddleware();
         $result = $middleware->process($request->reveal(), $delegate->reveal());